requirements.txt
# FastAPI and web framework
fastapi==0.115.12
uvicorn==0.35.0
pydantic==2.10.6

# LangChain and AI
langchain-community==0.3.26
langchain-huggingface>=0.1.0,<0.2.0
sentence-transformers==5.0.0

# Database (using psycopg2 for compatibility)
psycopg2-binary==2.9.10

# AWS
boto3>=1.34.0,<2.0.0

# Utilities
python-dotenv==1.1.1

# ML dependencies
torch --index-url https://download.pytorch.org/whl/cpu
numpy
transformers
huggingface-hub